backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
pillow==10.4.0
pymongo==4.10.1
python-dotenv==1.0.1
numpy>=1.24,<2.2
werkzeug==3.0.1
scikit-learn==1.5.2
matplotlib==3.9.2
requests==2.32.3
seaborn==0.13.2

